refactor: inline completion provider
