+++ packages/ai-native/src/browser/contrib/intelligent-completions/intelligent-completions.ts
@@ -1,12 +1,7 @@
-export interface IIntelligentCompletionsResult {
-  readonly items: IIntelligentCompletionItem[];
-}
+import { isUndefined } from '@opensumi/ide-core-common';
+import { InlineCompletion } from '@opensumi/ide-monaco';
 
-export interface IIntelligentCompletionItem {
-  /**
-   * 补全的内容
-   */
-  content: string;
+export interface IIntelligentCompletionItem extends InlineCompletion {
   /**
    * 表示当前行之前需要再补全几行
    * 最大值为 3
@@ -17,8 +12,14 @@ export interface IIntelligentCompletionItem {
    * 最大值为 3
    */
   belowRadius?: number;
+}
+
+export interface IIntelligentCompletionsResult<T = any> {
+  readonly items: IIntelligentCompletionItem[];
   /**
    * 定义的额外信息
    */
-  extra?: any;
+  extra?: T;
 }
+
+export const isMultiLineCompletion = (item: IIntelligentCompletionItem) => !isUndefined(item.aboveRadius) || !isUndefined(item.belowRadius);
