+++ packages/ai-native/src/browser/types.ts
@@ -209,11 +209,6 @@ export const AINativeCoreContribution = Symbol('AINativeCoreContribution');
 
 export interface AINativeCoreContribution {
   /**
-   * 通过中间件扩展部分 ai 能力
-   */
-  middleware?: IAIMiddleware;
-
-  /**
    * 注册 inline chat 相关功能
    * @param registry: IInlineChatFeatureRegistry
    */
@@ -256,6 +251,9 @@ export interface IChatAgentViewService {
   getChatComponentDeferred(componentId: string): Deferred<IChatComponentConfig> | null;
 }
 
+/**
+ * @deprecated use registerIntelligentCompletionProvider API
+ */
 export type IProvideInlineCompletionsSignature = (
   this: void,
   model: ITextModel,
@@ -265,6 +263,9 @@ export type IProvideInlineCompletionsSignature = (
   requestOption: IAICompletionOption,
 ) => MaybePromise<IAICompletionResultModel | null>;
 
+/**
+ * @deprecated use registerIntelligentCompletionProvider API
+ */
 export interface IAIMiddleware {
   language?: {
     provideInlineCompletions?: IProvideInlineCompletionsSignature;
